refactor: extract main.go logic into separate packages
