+++ .gitignore
@@ -23,3 +23,7 @@ vendor/
 # IDEs directories
 .idea
 .vscode
+
+# Binaries
+agent
+server
